+++ modules/upload/filetype.go
@@ -2,15 +2,30 @@
 // Use of this source code is governed by a MIT-style
 // license that can be found in the LICENSE file.
 
-package utils
+package upload
 
 import (
+	"fmt"
 	"net/http"
 	"strings"
 
-	"code.gitea.io/gitea/models"
 	"code.gitea.io/gitea/modules/log"
 )
+
+// ErrFileTypeForbidden not allowed file type error
+type ErrFileTypeForbidden struct {
+	Type string
+}
+
+// IsErrFileTypeForbidden checks if an error is a ErrFileTypeForbidden.
+func IsErrFileTypeForbidden(err error) bool {
+	_, ok := err.(ErrFileTypeForbidden)
+	return ok
+}
+
+func (err ErrFileTypeForbidden) Error() string {
+	return fmt.Sprintf("File type is not allowed: %s", err.Type)
+}
 
 // VerifyAllowedContentType validates a file is allowed to be uploaded.
 func VerifyAllowedContentType(buf []byte, allowedTypes []string) error {
@@ -27,7 +42,7 @@ func VerifyAllowedContentType(buf []byte, allowedTypes []string) error {
 
 	if !allowed {
 		log.Info("Attachment with type %s blocked from upload", fileType)
-		return models.ErrFileTypeForbidden{Type: fileType}
+		return ErrFileTypeForbidden{Type: fileType}
 	}
 
 	return nil
